Better recursive implementation
Now that the work buffer is stored as an immutable Static Vector, we
don't have to worry about displacing the elements and moving them
back at the end.  This lets us write a non-mutating recursive
evaluation algorithm, which is
a) easier to reason about
b) not as hard on the compiler, because we can shorten the length of
the SVector needed.