`AdaptiveSparseGrid` isn't a`Function`
We were setting these up as subtypes of functions, but this was
overriding the custom show method.

In addition, for some reason this was causing bugs in vs-code (though I
have no idea why...)

Unfortunately this is technically a breaking change (though it's pretty
minor).  In accordance with Semver, we have bumped the version to
0.2.0